[Upgrade] set dtype to X, y in split/utils.py

+++ dojo/split/utils.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 def convert_assert(X, y):
-    X, y = np.array(X), np.array(y)
+    X, y = np.array(X, dtype=np.float32), np.array(y, dtype=np.float32)
     assert X.shape[0] == y.shape[0]
 
     return X, y
